+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [3.20.3]
+
+- Add Huawei Cloud MaaS Provider (Thanks @ddling!)
+- Add Cerebras Qwen 3 235B instruct model (Thanks @kevint-cerebras!)
+- Add DeepSeek R1 0528 support under Hugging Face (Thanks @0ne0rZer0!)
+- Fix Global Rules directory documentation for Linux/WSL systems
+- Fix token counting when using VSCode LM API provider
+- Fix input field stealing focus issue by only focusing on visible and active editor panels
+- Fix duplicate tool registration for claude4-experimental
+- Trim input value for URL fields
+
 ## [3.20.2]
 
 - Fixed issue with sap ai core client credentials storage